Move code to extra file
